data for presentation

+++ src/components/Gallery/GalleryCard.js
@@ -1,26 +1,46 @@
 import React from "react";
 import { Link } from "react-router-dom"
-import { Card } from 'react-bootstrap'
+import { Card, Carousel } from 'react-bootstrap'
 import { CardDeck} from 'react-bootstrap'
 import { Button } from 'react-bootstrap'
 import { Figure } from 'react-bootstrap'
 
 const GalleryCard = props => {
+  const currentUser = parseInt(sessionStorage.getItem("activeUser"))
+  if (props.picture.userId == currentUser) {
     return (
-        <Figure
-        max width={171}
-        max height={180}>
-  <Figure.Image
-    src={props.picture.url}
-    
-  />
-  <Figure.Caption>
-    {props.picture.caption}
-  </Figure.Caption>
-  <Figure.Caption>uploaded by: {props.picture.username}</Figure.Caption>
-  <Button className="btn" variant="outline-primary" type="button" onClick={() => props.deletePicture(props.picture.id)}>Delete</ Button>
-</Figure>
+      <Card>
+      <Card.Img variant="top" src={props.picture.url} 
+      width={200}
+      height={200}/>
+      <Card.Body>
+        <Card.Text>
+          {props.picture.caption}
+        </Card.Text>
+        <Card.Text>
+         Uploaded by: {props.picture.username}
+        </Card.Text>
+        <Button className="btn" variant="outline-primary" type="button" onClick={() => props.deletePicture(props.picture.id)}>Delete</ Button>
+
+      </Card.Body>
+    </Card>
     )
-}
+} else {
+  return (
+    <Card>
+    <Card.Img variant="top" src={props.picture.url} 
+    width={100}
+    height={180}/>
+    <Card.Body>
+      <Card.Text>
+        {props.picture.caption}
+      </Card.Text>
+      <Card.Text>
+       Uploaded by: {props.picture.username}
+      </Card.Text>
+    </Card.Body>
+  </Card>
+  )
+}}
 
 export default GalleryCard
